cmake: check if correct libvmi is supplied
The fetchContent way of retrieving libvmi specifies the exact branch to
retrieve; with find_package, it is possible that an incorrect version of
libvmi is passed from vcpkg/nix/etc. This commit adds a check that
ensure the functions GDATA added to their libvmi fork are present. The
build would fail later, but this explicit error clarifies the issue.